float -> double

+++ src/geometrylib/intersects.cpp
@@ -4,16 +4,16 @@
 #include <string>
 using namespace std;
 int intersects(
-        float firstX,
-        float firstY,
-        float firstR,
-        float secondX,
-        float secondY,
-        float secondR)
+        double firstX,
+        double firstY,
+        double firstR,
+        double secondX,
+        double secondY,
+        double secondR)
 {
-    float distanceX = abs(firstX - secondX);
-    float distanceY = abs(firstY - secondY);
-    float distanceObs = sqrt((distanceX * distanceX) + (distanceY * distanceY));
+    double distanceX = abs(firstX - secondX);
+    double distanceY = abs(firstY - secondY);
+    double distanceObs = sqrt((distanceX * distanceX) + (distanceY * distanceY));
     if (distanceObs > firstR + secondR || distanceObs < abs(firstR - secondR)) {
         return 0;
     } else if (distanceObs == firstR + secondR) {
