+++ src/geometrylib/square.cpp
@@ -2,7 +2,7 @@
 #include <geometrylib/square.h>
 #define _USE_MATH_DEFINES
 #include <stdio.h>
-float calculate_square(float rad)
+double calculate_square(double rad)
 {
     return rad * rad * M_PI;
 }
